fix: implement graceful shutdown, non-blocking startup, and gRPC optimization
- Migrated bot and api startup to goroutines to prevent main thread blocking.
- Added signal notification for graceful system termination.
- Optimized xray client to use persistent gRPC connections.
- Finalized migration to the StatsProvider interface within the API layer.

+++ internal/models/stats.go
@@ -8,6 +8,6 @@ type StatsProvider interface {
 
 type UserStats struct {
 	Email string `json:"email"`
-	Downlink uint64 `json:"downlink"`
-	Uplink uint64 `json:"uplink"`
+	Downlink float64 `json:"downlink"`
+	Uplink float64 `json:"uplink"`
 }
